login
login optimalizalva, logout done backenden is

gergo ordenaresaga javitva

+++ backend/database/migrations/2023_02_28_124910_create_admin_codes_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('admin_codes', function (Blueprint $table) {
             $table->id();
             $table->string('roles', 20);
-            $table->bigInteger('code');
+            $table->smallInteger('code');
             $table->timestamps();
         });
     }
